Code for ticket #601 now compiles again.
git-svn-id: https://svn.code.sf.net/p/numere/source/branches/ticket_601@962 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ kernel/core/plotting/plotting.cpp
@@ -4135,7 +4135,7 @@ void Plot::getValuesFromData(DataAccessParser& _accessParser, size_t i, const st
         // if the current plot is a boxplot
         if (_pData.getBoxplot())
         {
-            vector<long long int> vJ = _idx.col.getVector();
+            vector<int> vJ = _idx.col.getVector();
             vJ.insert(vJ.begin(), -1);
             _idx.col = vJ;
         }
